:recycle: Navigation re-render 방지
throttle와 ref값 이용
