+++ src/components/Details/Navigation/index.ts
@@ -1,0 +1,2 @@
+export { DetialNavigation } from "./DetialNavigation";
+export { NavigationItem } from "./NavigationItem";
