+++ src/components/Details/index.ts
@@ -1,5 +1,4 @@
 export { DetailCard } from "./DetailCard";
 export { DetailContainer } from "./DetailContainer";
 export { DetailHeader } from "./DetailHeader/DetailHeader";
-export { DetialNavigation } from "./DetialNavigation";
 export { LeftHeader } from "./LeftHeader";
